added $env:OVS_RUNDIR
After installing OVS on windows, the environment is updated with a few OVS specific variables. These variables are needed by the userspace tools to determine its running environment. This variable is manually added to allow the ovs userspace tools to work without having to propagate the new environment to the current running hook.