Added s2d-health-check relation
In the hyper-converged deployments, when cluster nodes go down, the
Windows fail-over cluster will set the CSV's to offline and they need
to be manually brought back online when nodes come up again. We need
to trigger a health check for the S2D cluster whenever nodes boot up
for the first time. The s2d-proxy charm will make sure that all the
CSVs are properly started and mounted.